+++ changelog.md
@@ -4,9 +4,11 @@
 
 ### Upcoming
 ### Additions
+* Added Hud Infos
+  * A new Client Feature to display infos on the hud as long as you have the items
+* Fix for Villager Comfort: villagers that never sleept would not get comfort
 * Added a new StackMaxDamageEvent to change item stacks durability
 * Added a new EnchantmentBonusEfficiencyEvent to change efficiency before the other modifiers
-* Fix for Villager Comfort: villagers that never sleept would not get comfort
 
 ### Changes
 * Heavily increased animal spawn caps